requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
snowflake-connector-python[pandas]>=3.0.0